Add convolution and change working img to np.float

+++ imp/io.py
@@ -4,12 +4,12 @@ import numpy as np
 
 def read(fp):
     '''Open image with Pillow lib and return it as numpy array.'''
-    return np.asarray(Image.open(fp))
+    return np.asarray(Image.open(fp), dtype=np.float32)
 
 def write(img, fp):
     '''Write array as image to fp.'''
     try:
-        Image.fromarray(img).save(fp)
+        Image.fromarray(img.astype(np.uint8)).save(fp)
     except (KeyError, IOError) as e:
         print(e)
         return False
